[Update] NaverMapOptions 추가(Flutter side)
- 지도 초기값 설정 클래스인 NaverMapOptions 추가.
- 위에 해당하는 상수를 NaverMap 클래스에 추가.
- NaverMap 호출시 플랫폼에 NaverMapOptions를 전달하도록 변경.
- NaverMap 관련 UI enum 혹은 상수를 가질 ui.dart 파일 추가.
- Platform의존적인 상수를 가질 platform_constants.dart 파일 추가.
- 모든 zoom의 자료형을  double로 변경.

+++ lib/src/camera.dart
@@ -20,12 +20,12 @@ class CameraPosition {
   /// 줌 레벨.
   ///
   /// 이 값이 증가할수록 축척이 증가합니다.
-  final int zoom;
+  final double zoom;
 
   /// 카메라 위치에 관한 모든 요소를 지정해 객체를 생성합니다.
   CameraPosition(
     this.target, {
-    int zoom = 0,
+    double zoom = 0,
     double bearing = 0,
     double tilt = 0,
   })  : zoom = Geometry.clamp(zoom, NaverMap.minimumZoom, NaverMap.maximumZoom),
